requirements.txt
streamlit
PyPDF2
google.generativeai
python-dotenv                  # pip install -r requirement.txt 


# code is in my github